api/products part-2.1

+++ internal/product/dto.go
@@ -1,8 +1,8 @@
 package product
 
 type CreateProductRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Price       int64  `json:"price"`
-	Stock       int    `json:"stock"`
+	Name        string  `json:"name"`
+	Description string  `json:"description"`
+	Price       float64 `json:"price"`
+	Stock       int     `json:"stock"`
 }
